+++ Magic/src/main/java/com/elmakers/mine/bukkit/wand/Wand.java
@@ -2476,6 +2476,10 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
         List<String> lore = new ArrayList<>();
         for (String line : loreTemplate) {
             if (line == null) line = "";
+            if (mage != null) {
+                line = mage.parameterizeMessage(line);
+                line = mage.parameterizeAttributes(line);
+            }
             if (line.startsWith("$")) {
                 switch (line) {
                     case "$description":